main_app/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Cache
redis==5.1.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
